pyproject.toml
[project]
name = "rag-access-control"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.12",
    "google-api-python-client>=2.172.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.21",
    "numpy>=2.3.0",
    "openai>=1.85.0",
    "pandas>=2.3.0",
    "pinecone>=7.0.2",
    "pinecone-client>=6.0.0",
    "python-dotenv>=1.1.0",
    "slack-sdk>=3.35.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.115.9",
    "pydantic>=2.11.5",
    "uvicorn>=0.34.3",
]
jupyter = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "mypy>=1.16.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
    "pytest-cov>=6.1.1",
]
